Final updates to the main branch. (#97)
Things done:
- fixed alkos cli
- fixed some actions
- fixed compile warnings

---------

+++ alkos/kernel/test/string_test.cpp
@@ -15,7 +15,7 @@ class StringTest : public TestGroupBase
 
 TEST_F(StringTest, Strlen)
 {
-    R_ASSERT_EQ(13, strlen("Hello, World!"));
+    R_ASSERT_EQ(static_cast<size_t>(13), strlen("Hello, World!"));
     R_ASSERT_ZERO(strlen(""));
 }
